Generative/requirements.txt
# MARGDARSHAN - AI-Powered Career Guidance Platform
# Python Dependencies

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# AI and Machine Learning
google-generativeai==0.3.2
google-cloud-aiplatform==1.38.0
google-cloud-firestore==2.13.1
vertexai==1.38.0

# HTTP Client and Async Support
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Data Processing and Validation
pydantic==2.5.0
pydantic-settings==2.1.0
pandas==2.1.4
numpy==1.25.2

# Environment and Configuration
python-dotenv==1.0.0
python-multipart==0.0.6

# CORS and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Database Support (if needed)
sqlalchemy==2.0.23
alembic==1.13.1

# Testing Framework
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx==0.25.2

# Development Tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Optional AI Services
# Uncomment if using these services
# transformers==4.36.0
# torch==2.1.1
# openai==1.3.8
# anthropic==0.7.8

# Deployment and Production
gunicorn==21.2.0
python-multipart==0.0.6

# Monitoring and Logging
structlog==23.2.0
loguru==0.7.2

# Date and Time Handling
python-dateutil==2.8.2

# JSON Processing
orjson==3.9.10

# Regular Expressions (enhanced)
regex==2023.10.3

# Performance and Caching
redis==5.0.1
cachetools==5.3.2

# Background Tasks
celery==5.3.4

# File Handling
python-magic==0.4.27

# Email Support (if needed)
emails==0.6

# Template Engine (if needed)
jinja2==3.1.2

# Additional Utilities
python-slugify==8.0.1
bcrypt==4.1.2

# YouTube API Integration
# Using custom endpoint: https://abhi-api.vercel.app/api/search/yts

# Google Books API Integration
# Using official Google Books API: https://www.googleapis.com/books/v1/volumes# Using official Google Books API: https://www.googleapis.com/books/v1/volumes